+++ KalikoCMS.Admin/Admin/Content/EditPage.aspx.designer.cs
@@ -121,6 +121,24 @@ namespace KalikoCMS.Admin.Content {
         protected global::KalikoCMS.Admin.Content.PropertyType.NumericPropertyEditor SortOrder;
         
         /// <summary>
+        /// ChildSortOrder control.
+        /// </summary>
+        /// <remarks>
+        /// Auto-generated field.
+        /// To modify move field declaration from designer file to code-behind file.
+        /// </remarks>
+        protected global::System.Web.UI.WebControls.DropDownList ChildSortOrder;
+        
+        /// <summary>
+        /// ChildSortDirection control.
+        /// </summary>
+        /// <remarks>
+        /// Auto-generated field.
+        /// To modify move field declaration from designer file to code-behind file.
+        /// </remarks>
+        protected global::System.Web.UI.WebControls.DropDownList ChildSortDirection;
+        
+        /// <summary>
         /// EditControls control.
         /// </summary>
         /// <remarks>