+++ ReleaseNotes.md
@@ -5,6 +5,9 @@
 * Added option to set visibility in sitemaps when editing a page
 * Added page versioning
 * Added page published event
+* Added child sort order to pages
+* Added default child sort order to pagetypes
+* Get children now returns sorted list based on parent page settings
 * Divided startup code in pre and post CMS initialization
 * Improved custom routing in MVC provider
 * Added function to fetch all pages based on a predicate filter